chore(lint): ruff B/RUF 導入で検出したバックログを機械的に解消 (plan 024)
ruff select に B（flake8-bugbear）/ RUF を追加した際に検出された
約190件のバックログを解消した。RUF100（unused-noqa）/ RUF022（dunder-all
ソート）等は --fix で自動修正、B904（except 内 raise から from 欠落）/
RUF013（implicit Optional）/ RUF012（mutable class default → ClassVar）/
RUF005（list 結合 → unpacking）/ B905（zip strict 未指定）/ B007（未使用
ループ変数）等は個別に手動修正した。B905 は各 zip 呼び出しについて
呼び出し元のデータ構造から同長が自明なものだけ strict=True にし、
そうでないもの（呼び出し側が意図的に非同長を渡すケースがある等）は
strict=False（zip 既定と同じ挙動）にした。

+++ src/youtube_automation/utils/progress.py
@@ -85,7 +85,7 @@ def format_eta(seconds: float | None) -> str:
     """
     if seconds is None or seconds <= 0:
         return "--"
-    total = int(round(seconds))
+    total = round(seconds)
     if total < 60:
         return f"≈{total}s"
     m, s = divmod(total, 60)
@@ -201,7 +201,7 @@ def render_progress_line(
     if expected_total and expected_total > 0:
         progress = estimate_progress(elapsed, expected_total)
         eta = estimate_eta(elapsed, expected_total)
-        pct = int(round(progress * 100))
+        pct = round(progress * 100)
         eta_fmt = format_eta(eta)
         return f"{frame} {label}... {elapsed_fmt} (≈{pct}%, ETA {eta_fmt})"
     return f"{frame} {label}... {elapsed_fmt}"
